requirements.txt
pytest=6.2.1
assertpy=1.1
SQLAlchemy=1.3.20
requests=2.24.0
loguru=0.5.3
PyYAML=5.3.1
jsonschema=3.2.0
Jinja2=2.11.2
urllib3=1.25.10
colorama=0.4.3
colorlog=4.2.1
jsonpath=0.82